correct automation using mkdocs-gen-files
