+++ build_env.yml
@@ -32,6 +32,7 @@ dependencies:
   - xz=5.2.6
   - yaml=0.2.5
   - pip:
+      - mkdocs-gen-files==0.5.0
       - mkdocs-glightbox==0.3.7
       - mkdocs-git-revision-date-localized-plugin==1.2.6
       - mkdocs-redirects==1.2.1
